+++ lib/routes.dart
@@ -837,6 +837,7 @@ class AppRouterDelegate extends RouterDelegate<RouteConfiguration>
               chatService,
               myUserService,
               Get.find(),
+              Get.find(),
             ));
 
             deps.put(ChatWorker(chatService, myUserService, Get.find()));